move mock keymanager to its own testonly package
