+++ validator/client/BUILD.bazel
@@ -129,7 +129,7 @@ go_test(
         "//validator/db/testing:go_default_library",
         "//validator/graffiti:go_default_library",
         "//validator/keymanager/derived:go_default_library",
-        "//validator/keymanager/remote:go_default_library",
+        "//validator/keymanager/remote/mock:go_default_library",
         "//validator/slashing-protection-history:go_default_library",
         "//validator/testing:go_default_library",
         "@com_github_golang_mock//gomock:go_default_library",
