+++ validator/keymanager/remote/mock/BUILD.bazel
@@ -1,0 +1,15 @@
+load("@prysm//tools/go:def.bzl", "go_library")
+
+go_library(
+    name = "go_default_library",
+    testonly = True,
+    srcs = ["mock_keymanager.go"],
+    importpath = "github.com/prysmaticlabs/prysm/validator/keymanager/remote/mock",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//async/event:go_default_library",
+        "//config/fieldparams:go_default_library",
+        "//crypto/bls:go_default_library",
+        "//proto/prysm/v1alpha1/validator-client:go_default_library",
+    ],
+)
