+++ validator/keymanager/remote/mock/mock_keymanager.go
@@ -1,4 +1,4 @@
-package remote
+package mock
 
 import (
 	"context"
